add empty frontend filter

+++ service/static/js/rest_api.js
@@ -151,24 +151,27 @@ $(function () {
 
     $("#disable-btn").click(function () {
         var product_id = $("#product_id").val();
-
-        var ajax = $.ajax({
+        if(!product_id.trim().length) {
+            clear_form_data()
+            flash_message("Product Id should not be empty!")
+        } else {
+            var ajax = $.ajax({
                 type: "PUT",
                 url: "/inventory/" + product_id + "/disable",
                 contentType: "application/json",
                 data:''
             })
 
-        ajax.done(function(res){
-            update_form_data(res)
-            flash_message("Product " + product_id +" has been Disabled!")
-            list_all_inventories()
-        });
+            ajax.done(function(res){
+                update_form_data(res)
+                flash_message("Product " + product_id +" has been Disabled!")
+                list_all_inventories()
+            });
 
-        ajax.fail(function(res){
-            flash_message(res.responseJSON.message)
-        });
-
+            ajax.fail(function(res){
+                flash_message(res.responseJSON.message)
+            });
+        }
     });
 
     // **********************************************
@@ -215,23 +218,27 @@ $(function () {
         $("#search_results").empty();
         var inventory_id = $("#inventory_id").val();
 
-        var ajax = $.ajax({
-            type: "GET",
-            url: "/inventory/" + inventory_id,
-            contentType: "application/json",
-            data: ''
-        })
-
-        ajax.done(function(res){
-            update_form_data(res)
-            flash_message("Inventory "+ inventory_id +" has been Retrieved!")
-        });
-
-        ajax.fail(function(res){
+        if(!inventory_id.trim().length) {
             clear_form_data()
-            flash_message(res.responseJSON.message)
-        });
+            flash_message("Inventory Id should not be empty!")
+        } else {
+            var ajax = $.ajax({
+                type: "GET",
+                url: "/inventory/" + inventory_id,
+                contentType: "application/json",
+                data: ''
+            })
 
+            ajax.done(function(res){
+                update_form_data(res)
+                flash_message("Inventory "+ inventory_id +" has been Retrieved!")
+            });
+
+            ajax.fail(function(res){
+                clear_form_data()
+                flash_message(res.responseJSON.message)
+            });
+        }
     });
 
     // ****************************************
@@ -241,22 +248,27 @@ $(function () {
     $("#delete-btn").click(function () {
         var inventory_id = $("#inventory_id").val();
 
-        var ajax = $.ajax({
-            type: "DELETE",
-            url: "/inventory/" + inventory_id,
-            contentType: "application/json",
-            data: '',
-        })
-
-        ajax.done(function(res){
+        if(!inventory_id.trim().length) {
             clear_form_data()
-            flash_message("Inventory "+ inventory_id +" has been Deleted!")
-            list_all_inventories()
-        });
+            flash_message("Inventory Id should not be empty!")
+        } else {
+            var ajax = $.ajax({
+                type: "DELETE",
+                url: "/inventory/" + inventory_id,
+                contentType: "application/json",
+                data: '',
+            })
 
-        ajax.fail(function(res){
-            flash_message("Server error!")
-        });
+            ajax.done(function(res){
+                clear_form_data()
+                flash_message("Inventory "+ inventory_id +" has been Deleted!")
+                list_all_inventories()
+            });
+
+            ajax.fail(function(res){
+                flash_message(res.responseJSON.message)
+            });
+        }
     });
 
     // ****************************************
